Move feature config check upfront and extract a common method
